+++ application/core/XML_Model.php
@@ -46,27 +46,6 @@ class XML_Model extends Memory_Model
 	 */
 	protected function load()
 	{
-
-		if (($tasks = simplexml_load_file($this->_origin)) !== FALSE)
-		{
-			foreach ($tasks as $task) {
-				$record = new stdClass();
-				$record->id = (int) $task->id;
-				$record->task = (string) $task->task;
-				$record->priority = (int) $task->priority;
-				$record->size = (int) $task->size;
-				$record->group = (int) $task->group;
-				$record->deadline = (string) $task->deadline;
-				$record->status = (int) $task->status;
-				$record->flag = (int) $task->flag;
-
-				$this->_data[$record->id] = $record;
-			}
-		}
-
-		// rebuild the keys table
-		$this->reindex();
-
 		if (file_exists(realpath($this->_origin))) {
 
 		    $this->xml = simplexml_load_file(realpath($this->_origin));